backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Docker SDK
docker==7.0.0
docker[tls]==7.0.0

# File handling
python-multipart==0.0.6

# Development (optional)
# pytest==7.4.4
# pytest-asyncio==0.23.3
# black==24.1.1
# flake8==7.0.0